+++ lib/features/vendor_applications/vendor_applications_page.dart
@@ -105,8 +105,7 @@ class _VendorApplicationsPageState
         ? null
         : newestApplicationId;
     final int totalPages = (values.length / 10).ceil();
-    final int safePage =
-        totalPages == 0 ? 0 : page.clamp(0, totalPages - 1) as int;
+    final int safePage = totalPages == 0 ? 0 : page.clamp(0, totalPages - 1);
     return Column(
       children: [
         PageHeader(
